catch and report uncaught errors in wasm exported functions
Updates wasm exported functions to catch and report uncaught errors.

NOTE: the way we're generating wasmexports.zig is a little broken.  We're
not including the sources in its HASH so for now you need to manually
remove zig-cache if you want to regenerate wasmexports.zig.  You'll
need to do this if you change a wasm export function signature for
example.